requirements.txt
ccxt>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0
arch>=5.3.0
PyYAML>=6.0
scipy>=1.10.0
tqdm>=4.65.0